Issue-9750 - Extract array_length, array_reverse and array_sort functions from functions-array subcrate' s kernels and udf containers (#9751)
